% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fns.R
\name{plot_merges}
\alias{plot_merges}
\title{plot_merges}
\usage{
plot_merges(x, root_tree = FALSE)
}
\arguments{
\item{x}{Object of class \code{scl} obtained with \code{method = "exact"}.}

\item{root_tree}{If \code{TRUE}, tree leaves are connected to bottom of plot,
otherwise floating as determined by \link{plot.hclust}.}
}
\value{
Nothing (generates plot)
}
\description{
Plot dendrogram of merges for \code{scl} object with \code{method = "exact"}.
}
