% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{scl_cluster}
\alias{scl_cluster}
\title{scl_cluster}
\usage{
scl_cluster(xy, dmat, ncl, shortest = TRUE)
}
\arguments{
\item{xy}{Rectangular structure (matrix, data.frame, tibble), containing
coordinates of points to be clustered.}

\item{dmat}{Square structure (matrix, data.frame, tibble) containing
distances or equivalent metrics betwen all points in \code{xy}. If \code{xy}
has \code{n} rows, then \code{dat} must have \code{n} rows and \code{n}
columns.}

\item{ncl}{Desired number of clusters}

\item{shortest}{If \code{TRUE}, the \code{dmat} is interpreted as distances
such that lower values are preferentially selected; if \code{FALSE}, then
higher values of \code{dmat} are interpreted to indicate stronger
relationships, as is the case for example with covariances.}
}
\value{
A object of class \code{scl} with \code{tree} containing the
clustering scheme, and \code{xy} the original coordinate data of the
clustered points.
}
\description{
Cluster spatial data
}
\examples{
\dontrun{
xy <- matrix (runif (100), ncol = 2)
edges <- scl_edges (xy)
# add some fake data to the edges
edges \%<>\% dplyr::mutate (d = runif (nrow (.))) \%>\%
   dplyr::arrange (desc (d))
tree <- scl_spantree (edges) # plain tree; no components
ncl <- 8 # desired number of clusters/components
tree <- scl_cuttree (tree, edges, ncl = ncl) # tree with component numbers
}
}
