% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{scl_components}
\alias{scl_components}
\title{scl_components}
\usage{
scl_components(tree)
}
\arguments{
\item{tree}{result of \link{scl_spantree}}
}
\value{
Modified version of \code{tree}, with additional \code{comp} column
enumerating the component numbers
}
\description{
Get component vector of tree edges
}
