% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{scl_cuttree}
\alias{scl_cuttree}
\title{scl_cuttree}
\usage{
scl_cuttree(tree, edges, ncl)
}
\arguments{
\item{tree}{result of \link{scl_spantree}}

\item{edges}{A set of edges resulting from \link{scl_edges}, but with
additional data specifying edge weights, distances, or desired properties
from which to construct the tree}

\item{ncl}{Number of clusters or components into which tree is to be cut}
}
\value{
A tree
}
\description{
Cut a tree generated with \link{scl_spantree} into a specified number of
clusters or components
}
\examples{
xy <- matrix (runif (100), ncol = 2)
edges <- scl_edges (xy)
# add some fake data to the edges
edges \%<>\% dplyr::mutate (d = runif (nrow (.))) \%>\%
   dplyr::arrange (desc (d))
tree <- scl_spantree (edges)
ncl <- 8 # desired number of clusters/components
tree <- scl_cuttree (tree, edges, ncl = ncl)
}
