% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{scl_exact}
\alias{scl_exact}
\title{scl_exact}
\usage{
scl_exact(xy, dmat, ncl, method = "single")
}
\arguments{
\item{xy}{Rectangular structure (matrix, data.frame, tibble), containing
coordinates of points to be clustered.}

\item{dmat}{Square structure (matrix, data.frame, tibble) containing
distances or equivalent metrics betwen all points in \code{xy}. If \code{xy}
has \code{n} rows, then \code{dat} must have \code{n} rows and \code{n}
columns.}

\item{ncl}{Desired number of clusters}

\item{method}{One of \code{"single"}, \code{"average"}, or
\code{"complete"}; see \link{scl_redcap} for details.}
}
\value{
A object of class \code{scl} with \code{tree} containing the
clustering scheme, and \code{xy} the original coordinate data of the
clustered points. An additional component, \code{tree_rest}, enables the tree
to be re-cut to a different number of clusters via \link{scl_recluster},
rather than calculating clusters anew.
}
\description{
Exact spatially-constrained clustering.
}
\examples{
n <- 100
xy <- matrix (runif (2 * n), ncol = 2)
dmat <- matrix (runif (n ^ 2), ncol = n)
scl <- scl_exact (xy, dmat, ncl = 4)
}
