% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scl-full.R
\name{scl_full}
\alias{scl_full}
\title{scl_full}
\usage{
scl_full(xy, dmat, ncl, linkage = "single", shortest = TRUE, nnbs = 6L)
}
\arguments{
\item{xy}{Rectangular structure (matrix, data.frame, tibble), containing
coordinates of points to be clustered.}

\item{dmat}{Square structure (matrix, data.frame, tibble) containing
distances or equivalent metrics between all points in \code{xy}. If \code{xy}
has \code{n} rows, then \code{dat} must have \code{n} rows and \code{n}
columns.}

\item{ncl}{Desired number of clusters. See description of `ncl_iterate`
parameter for conditions under which actual number may be less than this
value.}

\item{linkage}{Either \code{"single"} or \code{"average"}. For covariance
clustering, use \code{"single"} with `shortest = FALSE`.}

\item{shortest}{If \code{TRUE}, the \code{dmat} is interpreted as distances
such that lower values are preferentially selected; if \code{FALSE}, then
higher values of \code{dmat} are interpreted to indicate stronger
relationships, as is the case for example with covariances.}

\item{nnbs}{Number of nearest neighbours to be used in calculating clustering
trees. Triangulation will be used if \code{nnbs <= 0}.}
}
\description{
Full spatially-constrained clustering.
}
\examples{
n <- 100
xy <- matrix (runif (2 * n), ncol = 2)
dmat <- matrix (runif (n^2), ncol = n)
scl <- scl_full (xy, dmat, ncl = 4)
}
\seealso{
Other clustering_fns: 
\code{\link{scl_recluster}()},
\code{\link{scl_redcap}()}
}
\concept{clustering_fns}
