% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fns.R
\name{scl_plot}
\alias{scl_plot}
\title{scl_plot}
\usage{
scl_plot(tree, xy)
}
\arguments{
\item{tree}{Spanning tree obtained from \link{scl_components}}

\item{xy}{Matrix of spatial coordinates of points indexed by \code{tree}.}
}
\value{
(Invisible) \link{ggplot2} object containing the plot
}
\description{
plot cluster groups
}
