% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{scl_recluster}
\alias{scl_recluster}
\title{scl_reccluster}
\usage{
scl_recluster(scl, ncl, shortest = TRUE)
}
\arguments{
\item{scl}{An \code{scl} object returned from \link{scl_redcap}.}

\item{ncl}{Desired number of clusters}

\item{shortest}{If \code{TRUE}, the \code{dmat} is interpreted as distances
such that lower values are preferentially selected; if \code{FALSE}, then
higher values of \code{dmat} are interpreted to indicate stronger
relationships, as is the case for example with covariances.}
}
\value{
Modified \code{scl} object in which \code{tree} is re-cut into
\code{ncl} clusters.
}
\description{
Re-cut a spatial cluster tree (\code{scl}) at a different number of clusters.
}
\examples{
n <- 100
xy <- matrix (runif (2 * n), ncol = 2)
dmat <- matrix (runif (n ^ 2), ncol = n)
scl <- scl_redcap (xy, dmat, ncl = 4)
plot (scl)
scl <- scl_recluster (scl, ncl = 5)
plot (scl)
}
