% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{scl_spantree}
\alias{scl_spantree}
\title{scl_spantree}
\usage{
scl_spantree(edges)
}
\arguments{
\item{edges}{A set of edges resulting from \link{scl_edges}, but with
additional data specifying edge weights, distances, or desired properties
from which to construct the tree}
}
\value{
A tree
}
\description{
Generate a spanning tree from a set of edges
}
\examples{
xy <- matrix (runif (100), ncol = 2)
edges <- scl_edges (xy)
# add some fake data to the edges
edges \%<>\% mutate (d = runif (nrow (.)), id = seq (nrow (.))) \%>\%
   arrange (desc (d))
tree <- scl_spantree (edges)
\dontrun{
# plot the tree
plot (xy, pch = 19)
with (edges, segments (xy [from, 1], xy [from, 2], xy [to, 1], xy [to, 2],
                       col = "gray", lty = 2))
with (tree, segments (xy [from, 1], xy [from, 2], xy [to, 1], xy [to, 2],
                      col = "red", lwd = 4))
}
}
