% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcluster-package.r
\docType{package}
\name{spatialcluster}
\alias{spatialcluster}
\title{spatialcluster.}
\description{
R port of redcap (Regionalization with dynamically constrained agglomerative
clustering and partitioning).
}
\concept{package}
