% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fns.R
\name{plot.scl}
\alias{plot.scl}
\title{plot.scl}
\usage{
\method{plot}{scl}(x, ...)
}
\arguments{
\item{x}{object to be plotted}

\item{...}{ignored here}
}
\description{
plot.scl
}
\examples{
n <- 20
xy <- matrix (runif (2 * n), ncol = 2)
dmat <- matrix (runif (n ^ 2), ncol = n)
scl <- scl_cluster (xy, dmat, ncl = 4, shortest = TRUE, full_order = FALSE)
plot (scl)
# Connect clusters according to highest (\\code{shortest = FALSE}) values of
# \\coce{dmat}:
scl <- scl_cluster (xy, dmat, ncl = 4, shortest = FALSE, full_order = FALSE)
plot (scl)
}
