% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fns.R
\name{plot.scl}
\alias{plot.scl}
\title{plot.scl}
\usage{
\method{plot}{scl}(x, ..., hull_alpha = 1)
}
\arguments{
\item{x}{object to be plotted}

\item{...}{ignored here}

\item{hull_alpha}{alpha value of (non-)convex hulls, with default generating
a convex hull, and smaller values generating concave hulls. (See
?alphashape::ashape for details).}
}
\description{
plot.scl
}
\examples{
set.seed (1)
n <- 100
xy <- matrix (runif (2 * n), ncol = 2)
dmat <- matrix (runif (n ^ 2), ncol = n)
scl <- scl_redcap (xy, dmat, ncl = 4)
plot (scl)
# Connect clusters according to highest (\code{shortest = FALSE}) values of
# \code{dmat}:
scl <- scl_redcap (xy, dmat, ncl = 4, shortest = FALSE, full_order = FALSE)
plot (scl)
}
\seealso{
Other plot_fns: 
\code{\link{plot_merges}()}
}
\concept{plot_fns}
