% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fns.R
\name{plot.scl}
\alias{plot.scl}
\title{plot.scl}
\usage{
\method{plot}{scl}(x, ..., convex = TRUE, hull_alpha = 0.1)
}
\arguments{
\item{x}{object to be plotted}

\item{...}{ignored here}

\item{convex}{Should hull be convex? If not, the \code{ashape} routine from
the \pkg{alphahull} package is used to generate non-convex hulls, generated
with the \code{hull_alpha} parameter}

\item{hull_alpha}{alpha value of non-convex hulls (see ?alphashape::ashape
for details).}
}
\description{
plot.scl
}
\examples{
n <- 100
xy <- matrix (runif (2 * n), ncol = 2)
dmat <- matrix (runif (n ^ 2), ncol = n)
scl <- scl_redcap (xy, dmat, ncl = 4)
plot (scl)
# Connect clusters according to highest (\code{shortest = FALSE}) values of
# \code{dmat}:
scl <- scl_redcap (xy, dmat, ncl = 4, shortest = FALSE, full_order = FALSE)
plot (scl)
}
