% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{scl_cluster}
\alias{scl_cluster}
\title{scl_cluster}
\usage{
scl_cluster(xy, dmat, ncl, shortest = TRUE)
}
\arguments{
\item{xy}{Rectangular structure (matrix, data.frame, tibble), containing
coordinates of points to be clustered.}

\item{dmat}{Square structure (matrix, data.frame, tibble) containing
distances or equivalent metrics betwen all points in \code{xy}. If \code{xy}
has \code{n} rows, then \code{dat} must have \code{n} rows and \code{n}
columns.}

\item{ncl}{Desired number of clusters}

\item{shortest}{If \code{TRUE}, the \code{dmat} is interpreted as distances
such that lower values are preferentially selected; if \code{FALSE}, then
higher values of \code{dmat} are interpreted to indicate stronger
relationships, as is the case for example with covariances.}
}
\value{
A object of class \code{scl} with \code{tree} containing the
clustering scheme, and \code{xy} the original coordinate data of the
clustered points. An additional component, \code{tree_rest}, enables the tree
to be re-cut to a different number of clusters via \link{scl_recluster},
rather than calculating clusters anew.
}
\description{
Cluster spatial data
}
\examples{
n <- 20
xy <- matrix (runif (2 * n), ncol = 2)
dmat <- matrix (runif (n ^ 2), ncol = n)
scl <- scl_cluster (xy, dmat, ncl = 4)
# Thos clusters will by default be constructed by connecting edges with the
# lowest (\\code{shortest}) values of \\code{dmat}, and will differ from
scl <- scl_cluster (xy, dmat, ncl = 4, shortest = FALSE)
}
