% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{scl_components}
\alias{scl_components}
\title{scl_components}
\usage{
scl_components(tree)
}
\arguments{
\item{tree}{result of \link{scl_spantree}}
}
\value{
Modified version of \code{tree}, with additional \code{comp} column
enumerating the component numbers
}
\description{
Get component vector of tree edges
}
\examples{
xy <- matrix (runif (100), ncol = 2)
edges <- scl_edges (xy)
# add some fake data to the edges
edges \%<>\% dplyr::mutate (d = runif (nrow (.))) \%>\%
   dplyr::arrange (desc (d))
tree <- scl_spantree (edges) # plain tree; no components
ncl <- 8 # desired number of clusters/components
tree <- scl_cuttree (tree, edges, ncl = ncl) # tree with component numbers
}
