% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{scl_edges}
\alias{scl_edges}
\title{scl_edges}
\usage{
scl_edges(xy)
}
\arguments{
\item{xy}{A matrix, data.frame, or other rectangular structure containing x
and y coordinates for triangulation}
}
\value{
A \pkg{tibble} of two columns indexing into \code{xy}, where each row
denotes the start and end of an edge
}
\description{
Generate triangulated edges between a set of input points
}
