% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{scl_exact}
\alias{scl_exact}
\title{scl_exact}
\usage{
scl_exact(xy, dmat, ncl, linkage = "single", shortest = TRUE,
  nnbs = 6)
}
\arguments{
\item{xy}{Rectangular structure (matrix, data.frame, tibble), containing
coordinates of points to be clustered.}

\item{dmat}{Square structure (matrix, data.frame, tibble) containing
distances or equivalent metrics betwen all points in \code{xy}. If \code{xy}
has \code{n} rows, then \code{dat} must have \code{n} rows and \code{n}
columns.}

\item{ncl}{Desired number of clusters}

\item{linkage}{One of \code{"single"}, \code{"average"}, or
\code{"complete"}; see Note.}

\item{shortest}{If \code{TRUE}, the \code{dmat} is interpreted as distances
such that lower values are preferentially selected; if \code{FALSE}, then
higher values of \code{dmat} are interpreted to indicate stronger
relationships, as is the case for example with covariances.}

\item{nnbs}{Number of nearest neighbours to be used in calculating clustering
trees. Triangulation will be used if \code{nnbs <= 0}.}
}
\description{
Exact spatially-constrained clustering.
}
\examples{
n <- 100
xy <- matrix (runif (2 * n), ncol = 2)
dmat <- matrix (runif (n ^ 2), ncol = n)
scl <- scl_exact (xy, dmat, ncl = 4)
}
