% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fns.R
\name{scl_plot}
\alias{scl_plot}
\title{scl_plot}
\usage{
scl_plot(tree, xy)
}
\arguments{
\item{tree}{Spanning tree obtained from \link{scl_cluster}}

\item{xy}{Matrix of spatial coordinates of points indexed by \code{tree}.}
}
\value{
(Invisible) \link{ggplot2} object containing the plot
}
\description{
plot cluster groups
}
\examples{
\dontrun{
xy <- matrix (runif (100), ncol = 2)
edges <- scl_edges (xy)
# add some fake data to the edges
edges \%<>\% dplyr::mutate (d = runif (nrow (.))) \%>\%
   dplyr::arrange (desc (d))
# get tree with component numbers
ncl <- 12 # desired number of clusters/components
tree <- scl_spantree (edges) \%>\%
    scl_cuttree (edges, ncl = ncl)
xy <- tibble::tibble (x = xy [, 1], y = xy [, 2])
g <- scl_plot (tree, xy)
}
}
