% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{scl_spantree}
\alias{scl_spantree}
\title{scl_spantree}
\usage{
scl_spantree(edges)
}
\arguments{
\item{edges}{A set of edges resulting from \link{scl_edges}, but with
additional data specifying edge weights, distances, or desired properties
from which to construct the tree}
}
\value{
A tree
}
\description{
Generate a spanning tree from a set of edges
}
