% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcluster-package.R
\docType{package}
\name{spatialcluster}
\alias{spatialcluster-package}
\alias{spatialcluster}
\title{spatialcluster.}
\description{
R port of redcap (Regionalization with dynamically constrained agglomerative
clustering and partitioning).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpadge/spatialcluster}
  \item Report bugs at \url{https://github.com/mpadge/spatialcluster/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\concept{package}
